Split Tests into their own files & add some - remove comments generated from vscode (in .vscode/) - adding tslint-eslint-rules to enforce indent(on save) - remove "@format" and apply it for every file for prettier - fix wrong indent in some test files - remove not needed reconnect - include tests/ for tslint & vscode to check - extra tsconfig for build that excludes test/ - set vscode to indent 2, instead of global - small style changes
- moving "should add" tests to their own file
- moving "select" tests to their own file
- moving "string Validator" tests to their own file
- moving "biguser" test to its own file
- moving "getClassForDocument" to its own file

- modify Improvemnts in README
- moved hooktests/* & renamed them
- adding "minlength" to the stringValidator tests
