+++ src/errors.ts
@@ -1,5 +1,3 @@
-/** @format */
-
 export class InvalidPropError extends Error {
   constructor(typeName: string, key: string) {
     super(`In property ${key}: ${typeName} is not a primitive type nor a Typegoose schema (Not extending it).`);
@@ -22,8 +20,8 @@ export class NoMetadataError extends Error {
   constructor(key: string) {
     super(
       `There is no metadata for the "${key}" property. ` +
-        'Check if emitDecoratorMetadata is enabled in tsconfig.json ' +
-        'or check if you\'ve declared a sub document\'s class after usage.'
+      'Check if emitDecoratorMetadata is enabled in tsconfig.json ' +
+      'or check if you\'ve declared a sub document\'s class after usage.'
     );
   }
 }
